Updated Code
So I made some minor changes to the code and added the secondary part of
the game. The game generates another ship at this time that we can place
on the number line as well and the text updates each time. When the
ships are placed in the right places as well they become un-clickable.