Score Board and Shot Counter
The Score works and carries over between games, but resets on page
refresh since it's only a variable being stored. I also added the number
of shots they have left and made it subtract/add shots when they are
working with the second ship. The adding shots part is unnecessary as of
right now since the game ends and asks if you want to restart when you
get it right, but we could use it later if we do extra stuff.